Simplify the `TimeSeries` code by removing the `Series` base class. Originally the base class was going to support time, timerange and index based series as three sub-classes. But it's much simpler to just keep the three different types of lists as members on the `TimeSeries` itself.
The TimeSeries may either have a list of `_times`, `_timeRanges` or `_indexes`. The rest of the code can now use the existence of those lists to perform operations that are dependent on the type of the series, rather the guess by if there's a string, list, or date. This means this (very simple) decision can often be made once and not per row, so this should be faster. Needs testing around TimeRange based TimeSeries still.

Also enables chrome dev tools custom formatting so that printing out a pond object doesn't drop you into the ugly immutable structures that back Pond.