Getting project ready for 0.7 release:  - Website now built with create-react-app  - Converted tests to use Jest (for CRA workflow)  - Cleaner API docs, more examples, many fixes
There's been quite a bit of reorganization to get the above in a position to work. While I was in there I did some renaming and project restructuring. Files are better named and following the python code, processors are organized in a sub-directory. Also moved into other subdirectories are io/ stuff (ins and outs to the pipelines), and a base/ directory which has utility code etc. This leaves the public API at the top level.

General API changes:
 - UnboundedIn is now Stream. That seems to make more sense when used.
 - BoundedIn is now just Bounded, but that's usually just a base class for something, so less user facing (unless you make a new Bounded source).

TimeSeries:
 - I also reworked the API for the pipeline wrapping functions in TimeSeries. They now all take a single options argument, which would contain the fieldSpec, window, reducer, etc within the options arg object. This makes it much more robust that having 6 arguments, some of which are optional, and some of which have default values
 - Renamed the `window` option in `TimeSeries.align()` to `period`, because that seemed to map better to what it is. (updated the missing data doc file)