[BREAKING CHANGE]
Adds the ability to subclass Events and TimeSeries

Builds out a proper class structure for Events, but adds a major breaking change:
  - `Event` is now the base class and will throw an error if instantiated directly, `TimeEvent` replaces `Event`.
  - This makes it clearer what the event is and how it is different than its siblings
  - Keeps static operations on the `Event` base class
  - Allows `TimeEvent`, `TimeRangeEvent` and `IndexedEvent` to be subclassed themselves (for Avro support, see below)
  - Simplifies the event code as well as several other places in the code

Avro support:
  - Adds the ability to specify a schema for `TimeRanges`, `Events` and `TimeSeries` via subclassing
  - `toAvro()` function uses the schema to build the Avro Buffer that could be sent across the wire or stored compactly and with validation
  - Constructor can now take a Buffer and do the conversion, as you'd expect.

Some of the above isn't complete. For one, TimeRangeEvent isn't done yet. Also, it would be nice if you could use a custom Event subclass in a TimeSeries, and then have the TimeSeries code use that for the schema specification. That wouldn't really provide a schema for the meta data though, so you'd still have to subclass. But it would be more unified.