The purpose of this change is to enable the `timeSeriesListReduce()` function to take a standard reducer function (`sum()`, `avg()`, etc) rather than a special event reducer (`Event.sum()`, `Event.avg()`, etc) which is a source of confusion and limits the reducers that can be used, as well as making it more complicated to implement your own reducer function because you would need to deal with fieldSpecs in each implementation.
In doing this we've removed the convenience functions that wrapped `timeSeriesListReduce()`, specifically `timeSeriesListSum()` and `timeSeriesListAvg()` in favor of less code and more composablity.

For example, here's the alternative to `timeSeriesListSum()`. Note the use of the `sum()` reducer function. Otherwise it's the same call.
```
const totalSeries = TimeSeries.timeSeriesListReduce({
    name: "totals",
    seriesList: [inTraffic, outTraffic],
    reducer: sum(),
    fieldSpec: [ "in", "out" ]
});
```

Similarly the little used (directly) `Event.sum()` and `Event.avg()` are gone. Use `Event.combine()` instead.