Extended to `Stream` constructor to accept another stream
This allows streams to be branched. In fact the Stream constructor accepts either of the two interfaces (`EventStream` or `KeyedCollectionStream`) in the form of a new base class `StreamInterface`. A `StreamInterface` holds just the reference back to the `Stream` itself. In this way, either interface can be passed into the `stream()` factory function and the nodes in that interface can be connected correctly to the new `Stream`.

Added a test for this.