Stream changes
The major change here is that a `StreamInterface` now holds onto the `tail` of the `Node` processing tree vs the `Stream` which holds the root of the tree. This means that streams can be properly branched directly from each other.

Also moved all the `Node` subclasses into their own file and did a pass over the docs.

Lib, docs and website rebuild.