fix: output of path invalid value in the case of null value
Previously when calculating rates if either the previous or current values were null pond would emit a console.log message saying the value was invalid or missing. In the case of nulls though it is better to just output a null rate. Nulls are often seen in timeseries when incomplete data is present.